so many fucking issues at graddle

+++ src/main/java/com/cinema/testcinema/dto/MovieDto.java
@@ -1,14 +1,14 @@
 package com.cinema.testcinema.dto;
 
 public class MovieDto {
-    private String title;
-    private int year;        // int, а не long
-    private String imdbId;
+    private String title;  // Игнорируется, если есть imdbId (берётся из OMDB)
+    private Long year;     // Теперь Long для consistency
+    private String imdbId; // Обязательно для автозаполнения
     private Long genreId;
 
     public MovieDto() {}
 
-    public MovieDto(String title, int year, String imdbId, Long genreId) {
+    public MovieDto(String title, Long year, String imdbId, Long genreId) {
         this.title = title;
         this.year = year;
         this.imdbId = imdbId;
@@ -19,8 +19,8 @@ public class MovieDto {
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
 
-    public int getYear() { return year; }      // исправлено
-    public void setYear(int year) { this.year = year; }
+    public Long getYear() { return year; }
+    public void setYear(Long year) { this.year = year; }
 
     public String getImdbId() { return imdbId; }
     public void setImdbId(String imdbId) { this.imdbId = imdbId; }
@@ -28,4 +28,3 @@ public class MovieDto {
     public Long getGenreId() { return genreId; }
     public void setGenreId(Long genreId) { this.genreId = genreId; }
 }
-
